requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
snowflake-snowpark-python